% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BANC}
\alias{BANC}
\title{Provisional template brain object for BANC}
\format{
An object of class \code{templatebrain} of dimension 262144 x 294912 x 7010.
}
\usage{
BANC
}
\description{
This is calibrated in nm. See \code{data-raw/BANC.R} for details
of how it was generated. It is still provisional pending synpase
cloud-based registrations.
}
\examples{
BANC
# Depends on nat
library(nat)
boundingbox(BANC)
}
\keyword{datasets}
