% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BANC.surf}
\alias{BANC.surf}
\title{Simplified tissue surface of BANC}
\format{
An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
BANC.surf
}
\description{
This is unsymmetrical and not normalized version of the mesh. It
is calibrated in nm. See \code{data-raw/BANC.R} for details of how it was
generated.
}
\examples{
\dontrun{
# Depends on nat
library(nat)
wire3d(BANC.surf, col='grey')
}
}
\keyword{datasets}
