% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{banc_lr_position}
\alias{banc_lr_position}
\title{Predict whether a point is on the left or right of the BANC dataset}
\usage{
banc_lr_position(x, units = c("nm", "microns", "raw"), group = FALSE)
}
\arguments{
\item{x}{An object from which xyzmatrix can extract points, calibrated in nm.}

\item{units}{Units for both input \emph{and} output data.}

\item{group}{Whether to return the mean displacement per neuron (when
\code{x} is a \code{neuronlist})}
}
\value{
A vector of point displacements (calibrated according to
\code{units}, nm is the default) where 0 is at the midline and positive
values are to the fly's right.
}
\description{
Predict whether a point is on the left or right of the BANC dataset
}
\details{
This is not perfect as it assumes that the X displacement from the
midline is a good indicator of LR displacement. This is generally true but
not infallible. Furthermore it will only be as good as the registration
used by \code{\link{mirror_banc}} (still variable).
}
\examples{
library(nat)
lrdiffs=banc_lr_position(xyzmatrix(BANC.surf))
\dontrun{
points3d(xyzmatrix(BANC.surf), col=ifelse(lrdiffs>0, 'green', 'red'))
}
}
\seealso{
\code{\link{mirror_banc}}
}
