% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudvolume.R
\name{dr_fanc}
\alias{dr_fanc}
\title{Print information about your FANC setup including tokens and python modules}
\usage{
dr_fanc()
}
\description{
Print information about your FANC setup including tokens and python modules
}
\examples{
\dontrun{
dr_fanc()
}
}
\seealso{
\code{\link{dr_fafbseg}}
}
