% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanc-package.R
\docType{package}
\name{fanc-package}
\alias{fanc}
\alias{fanc-package}
\title{fanc: Support Acccess to the Full Adult Nerve Cord (FANC) Dataset}
\description{
More about what it does (maybe more than one line) Use four
    spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://flyconnectome.github.io/fanc}
  \item \url{https://github.com/flyconnectome/fanc}
  \item Report bugs at \url{https://github.com/flyconnectome/fanc/issues}
}

}
\keyword{internal}
