% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FANC.surf}
\alias{FANC.surf}
\title{Simplified tissue surface of FANC}
\format{
An object of class \code{mesh3d} of length 3.
}
\usage{
FANC.surf
}
\description{
This is unsymmetrical and not normalized version of the mesh.
}
\examples{
\dontrun{
# Depends on nat
library(nat)
wire3d(FANC.surf)
}
}
\keyword{datasets}
