% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{fanc4to3}
\alias{fanc4to3}
\title{Convert XYZ locations between FANC4 (autoseg) and FANC3 (CATMAID)}
\usage{
fanc4to3(
  xyz,
  method = c("mapmany", "map1"),
  chunksize = 40000,
  swap = FALSE,
  ...
)
}
\arguments{
\item{xyz}{An Nx3 matrix of coordinates}

\item{method}{Whether to map many points at once (default) or just one}

\item{chunksize}{The number of points to send to the server when mapping many
points at once.}

\item{swap}{When \code{TRUE} applies the deformation field in the opposite
direction e.g. to give a coarse mapping of points FAFB->FlyWire. This is
wrong but may be useful.}

\item{...}{Additional arguments passed to \code{httr::GET}/\code{POST}
operation}
}
\value{

}
\description{
Convert XYZ locations between FANC4 (autoseg) and FANC3 (CATMAID)
}
\examples{
fanc3=xyzmatrix(cbind(194569.2, 470101.3, 117630))
fanc4=cbind(45224, 109317, 2614)*c(4.3,4.3,45)
fanc4to3(fanc4)-fanc3

# rather small error for the approx inverse in this case
fanc4to3(fanc3, swap=TRUE)-fanc4

# ... so reverse mapping works in this case, but not always
fanc_xyz2id(fanc4to3(fanc3, swap = T))
}
