% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{fanc_cave_client}
\alias{fanc_cave_client}
\alias{banc_cave_client}
\title{Low level access to FANC's CAVE annotation infrastructure}
\usage{
fanc_cave_client()

banc_cave_client()
}
\value{
A reticulate R object wrapping the python CAVEclient.
}
\description{
Low level access to FANC's CAVE annotation infrastructure
}
\examples{
\donttest{
fcc=fanc_cave_client()
tables=fcc$annotation$get_tables()
fcc$materialize$get_table_metadata(tables[1])
}
}
