% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetta-api.R
\name{fanc_change_log}
\alias{fanc_change_log}
\title{Fetch change log information for one or more neurons}
\usage{
fanc_change_log(x, ...)
}
\arguments{
\item{x}{One or more fanc ids in any format understandable by
\code{\link[fafbseg]{ngl_segments}}}

\item{...}{Additional argument passed to \code{\link{pbsapply}} and then on
to \code{\link{flywire_fetch}} (via the private \code{fanc_fetch}).}
}
\value{
A list containing information on the changes applied to a given body

\itemize{

\item \code{n_splits},\code{n_mergers} The number of splits/mergers applied
to the neuron

\item \code{user_info} A data.frame containing user ids or emails and the
number of splits/mergers

\item \code{operations_ids} An integer vector of operation ids

\item \code{past_ids} Previous ids associated with this body

}
}
\description{
Fetch change log information for one or more neurons
}
\examples{
\donttest{
fanc_change_log("648518346473954669")
}
}
