% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fanc_ids}
\alias{fanc_ids}
\title{Return a vector of FANC root ids from diverse inputs}
\usage{
fanc_ids(x, integer64 = NA)
}
\arguments{
\item{x}{A data.frame, URL or vector of ids}

\item{integer64}{Whether to return ids as \code{bit64::integer64} or
character vectors. Default value of NA leaves the ids unmodified.}
}
\value{
A vector of ids
}
\description{
Return a vector of FANC root ids from diverse inputs
}
\examples{
fanc_ids(data.frame(rootid="648518346474360770"))
}
\seealso{
Other fanc-ids: 
\code{\link{fanc_islatest}()},
\code{\link{fanc_latestid}()},
\code{\link{fanc_leaves}()},
\code{\link{fanc_rootid}()},
\code{\link{fanc_xyz2id}()}
}
\concept{fanc-ids}
