% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fanc_islatest}
\alias{fanc_islatest}
\title{Check if a FANC root id is up to date}
\usage{
fanc_islatest(x, timestamp = NULL, ...)
}
\arguments{
\item{x}{FlyWire rootids in any format understandable to
\code{\link[fafbseg]{ngl_segments}} including as \code{integer64}}

\item{timestamp}{(optional) argument to set an endpoint - edits after this
time will be ignored (see details).}

\item{...}{Additional arguments passed to \code{\link{flywire_islatest}}}
}
\description{
Check if a FANC root id is up to date
}
\examples{
fanc_islatest("648518346473954669")
}
\seealso{
\code{\link{fanc_latestid}}
}
