% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fanc_latestid}
\alias{fanc_latestid}
\title{Find the latest id for a FANC root id}
\usage{
fanc_latestid(
  rootid,
  timestamp = NULL,
  version = NULL,
  sample = 1000L,
  cloudvolume.url = NULL,
  Verbose = FALSE,
  ...
)
}
\arguments{
\item{rootid}{One ore more FlyWire rootids defining a segment (in any form
interpretable by \code{\link[fafbseg]{ngl_segments}})}

\item{timestamp}{An optional timestamp as a string or POSIXct, interpreted as
UTC when no timezone is specified.}

\item{version}{An optional CAVE materialisation version number. See details
and examples.}

\item{sample}{An absolute or fractional number of supervoxel ids to map to
rootids or \code{FALSE} (see details).}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL chooses the flywire production segmentation
dataset.}

\item{Verbose}{When set to \code{TRUE} prints information about what fraction
of}

\item{...}{Additional arguments passed to \code{\link{flywire_latestid}}}
}
\description{
Find the latest id for a FANC root id
}
\examples{
\dontrun{
fanc_latestid("648518346473954669")
# the latest materialisation version
fanc_latestid("648518346486614449", version='latest')
# up to the second
fanc_latestid("648518346486614449", timestamp='now')
}
}
\seealso{
\code{\link{fanc_islatest}}

Other fanc-ids: 
\code{\link{fanc_cellid_from_segid}()},
\code{\link{fanc_ids}()},
\code{\link{fanc_islatest}()},
\code{\link{fanc_leaves}()},
\code{\link{fanc_rootid}()},
\code{\link{fanc_xyz2id}()}
}
\concept{fanc-ids}
