% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fanc_leaves}
\alias{fanc_leaves}
\title{Find the supervoxel identifiers of a FANC neuron}
\usage{
fanc_leaves(x, integer64 = TRUE, ...)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{integer64}{Whether to return ids as integer64 type (the default, more
compact but a little fragile) rather than character (when \code{FALSE}).}

\item{...}{additional arguments passed to \code{\link{flywire_leaves}}}
}
\value{
A vector of supervoxel ids
}
\description{
Find the supervoxel identifiers of a FANC neuron
}
\examples{
\dontrun{
svids=fanc_leaves("648518346482929060")
head(svids)
}
}
\seealso{
\code{\link{flywire_leaves}}

Other fanc-ids: 
\code{\link{fanc_rootid}()},
\code{\link{fanc_xyz2id}()}
}
\concept{fanc-ids}
