% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners.R
\name{fanc_partner_summary}
\alias{fanc_partner_summary}
\title{Summarise the connectivity of FANC neurons}
\usage{
fanc_partner_summary(
  rootid,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0
)
}
\arguments{
\item{rootid}{Character vector specifying one or more flywire rootids. As a
convenience for \code{flywire_partner_summary} this argument is passed to
\code{\link[fafbseg]{ngl_segments}} allowing you to pass in}

\item{partners}{Whether to fetch input or output synapses or both.}

\item{threshold}{For \code{flywire_partner_summary} only return partners with
greater than this number of connections to the query neuron(s) (default of
0 returns all connections)}

\item{remove_autapses}{For \code{flywire_partner_summary} whether to remove
autapses (defaults to TRUE)}

\item{cleft.threshold}{A threshold for the cleft score calculated by Buhmann
et al 2019 (default 0, we have used 30-100 to increase specificity)}
}
\value{
a data.frame
}
\description{
Summarise the connectivity of FANC neurons
}
\examples{

}
