% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fanc_rootid}
\alias{fanc_rootid}
\title{Find the root identifier of a FANC neuron}
\usage{
fanc_rootid(x, integer64 = FALSE, ...)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{integer64}{Whether to return ids as integer64 type (more compact but a
little fragile) rather than character (default \code{FALSE}).}

\item{...}{Additional arguments passed to \code{pbapply::pbsapply} and
eventually to Python \code{cv$CloudVolume} object.}
}
\value{
A vector of root ids (by default character)
}
\description{
Find the root identifier of a FANC neuron
}
\examples{
\donttest{
fanc_rootid("73186243730767724")
}
}
\seealso{
\code{\link{flywire_rootid}}

Other fanc-ids: 
\code{\link{fanc_ids}()},
\code{\link{fanc_islatest}()},
\code{\link{fanc_latestid}()},
\code{\link{fanc_leaves}()},
\code{\link{fanc_xyz2id}()}
}
\concept{fanc-ids}
