% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{fanc_scene}
\alias{fanc_scene}
\title{Return a sample Neuroglancer scene URL for FANC dataset}
\usage{
fanc_scene(ids = NULL)
}
\arguments{
\item{ids}{A set of root ids to include in the scene.}
}
\value{
A character vector containing a single Neuroglancer URL.
}
\description{
Return a sample Neuroglancer scene URL for FANC dataset
}
\details{
See
\href{https://fanc-reconstruction.slack.com/archives/C01RZP5JH9C/p1616522511001900}{FANC
slack} for details.
}
\examples{
\donttest{
browseURL(fanc_scene())
}
}
