% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{fanc_voxdims}
\alias{fanc_voxdims}
\alias{fanc_nm2raw}
\alias{fanc_raw2nm}
\title{Handle raw and nm calibrated FANC coordinates}
\usage{
fanc_voxdims(url = choose_fanc(set = FALSE)[["fafbseg.sampleurl"]])

fanc_nm2raw(x, vd = fanc_voxdims())

fanc_raw2nm(x, vd = fanc_voxdims())
}
\arguments{
\item{url}{Optional neuroglancer URL containing voxel size. Defaults to
\code{getOption("fafbseg.sampleurl")} as set by
\code{\link{choose_fanc}}.}

\item{x}{3D coordinates in any form compatible with \code{\link{xyzmatrix}}}

\item{vd}{The voxel dimensions in nm. Expert use only. Normally found
automatically.}
}
\value{
For \code{fanc_voxdims} A 3-vector

for \code{fanc_raw2nm} and \code{fanc_nm2raw} an Nx3 matrix of
coordinates
}
\description{
\code{fanc_voxdims} returns the image voxel dimensions which
are normally used to scale between \bold{raw} and \bold{nm} coordinates.
}
\details{
relies on nat >= 1.10.4
}
\examples{
fanc_voxdims()
fanc_raw2nm(c(159144, 22192, 3560))
fanc_raw2nm('159144 22192 3560')
\dontrun{
fanc_nm2raw(clipr::read_clip())
}
}
