% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancr-package.R
\docType{package}
\name{fancr-package}
\alias{fancr}
\alias{fancr-package}
\title{fancr: Support Acccess to the Female Adult Nerve Cord (FANC) Dataset}
\description{
Access to the FANC EM dataset, focussing on support for
    autosegmentation. Organised as a thin wrapper around the fabseg
    package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/fancr}
  \item Report bugs at \url{https://github.com/flyconnectome/fancr/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
