% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\docType{data}
\name{FANCsym}
\alias{FANCsym}
\title{FANC symmetric template}
\format{
An object of class \code{templatebrain} of dimension 0 x 0 x 0.
}
\usage{
FANCsym
}
\description{
FANC symmetric template
}
\keyword{datasets}
