% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirror_banc_lm}
\alias{mirror_banc_lm}
\title{Provisional mirroring registration for BANC (lanmarks+thin plate splines)}
\format{
An object of class \code{tpsreg} of length 2.
}
\usage{
mirror_banc_lm
}
\description{
This is calibrated in nm. See \code{data-raw/mirror_banc.R} for
details of how it was generated. It is still provisional pending synpase
cloud-based registrations.
}
\examples{
utils::str(mirror_banc_lm)
}
\keyword{datasets}
