% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{mirror_fanc}
\alias{mirror_fanc}
\alias{symmetric_fanc}
\title{Mirror points or other 3D objects along the FANC midline}
\usage{
mirror_fanc(x, ...)

symmetric_fanc(x, mirror = FALSE, ...)
}
\arguments{
\item{x}{3D vertices (or object containing them) in FANC space Could be
\code{\link{neuron}}, \code{\link{neuronlist}}, \code{\link{hxsurf}} etc.}

\item{...}{additional arguments passed to \code{\link{xform}} and friends}

\item{mirror}{Whether to mirror across the midline when using
\code{symmetric_fanc}}
}
\value{
Transformed points/object
}
\description{
\code{symmetric_fanc} transforms neurons, surfaces and other
point data onto a symmetrised version of the FANC template brain,
optionally mirroring across the midline.
}
\details{
These registration functions depend on an installation of the CMTK
registration toolkit. See \code{\link[nat]{cmtk.bindir}} for details.
}
\examples{
\donttest{
# TODO
}

\dontrun{
FANC.surf.symm <- symmetric_fanc(FANC.surf)
# plot the two meshes: before and after
wire3d(FANC.surf, col='grey')
wire3d(FANC.surf.symm, col='red')
}
}
