% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{read_fanc_meshes}
\alias{read_fanc_meshes}
\title{Read one or more FANC neuron meshes}
\usage{
read_fanc_meshes(ids, savedir = NULL, format = c("ply", "obj"), ...)
}
\arguments{
\item{ids}{One or more root ids}

\item{savedir}{An optional location to save downloaded meshes. This acts as a
simple but effective cache since flywire neurons change id whenever they
are edited.}

\item{format}{whether to save meshes in Wavefront obj or Stanford poly
format. obj is the default but ply is a simpler and more compact format.}

\item{...}{Additional arguments passed to
\code{fafbseg::\link{read_cloudvolume_meshes}}}
}
\value{
A \code{\link[nat]{neuronlist}} containing one or more \code{mesh3d}
objects. See \code{nat::\link[nat]{read.neurons}} for details.
}
\description{
Read one or more FANC neuron meshes
}
\examples{
\donttest{
read_fanc_meshes("648518346482929060")
}
}
\seealso{
\code{fafbseg::\link{read_cloudvolume_meshes}}
}
