% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{transform_fanc2manc}
\alias{transform_fanc2manc}
\title{Transform FANC to MANC}
\usage{
transform_fanc2manc(x, inverse = F, ...)
}
\arguments{
\item{x}{an object to transform (calibrated in nm)}

\item{inverse}{boolean flag that says whether to swap the registration}

\item{...}{additional arguments passed to (passed on to \code{\link{xform}})}
}
\description{
transforms neurons, surfaces and other point data onto from FANC
to MANC space.
}
\examples{
\dontrun{
library(nat)
library(malevnc)
FANC.in.manc <- transform_fanc2manc(FANC.surf)
# plot MANC and FANC mesh for comparison
wire3d(malevnc::MANC.surf, col='grey', add=T)
wire3d(FANC.in.manc/1e3, col='blue',add=T)
}
}
