% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{fanc_cave_query}
\alias{fanc_cave_query}
\alias{banc_cave_query}
\title{Query FANC tables in the CAVE annotation system}
\usage{
fanc_cave_query(table, datastack_name = NULL, live = TRUE, ...)

banc_cave_query(table, datastack_name = NULL, live = TRUE, ...)
}
\arguments{
\item{table}{The name of the table (or view, see views section) to query}

\item{datastack_name}{An optional CAVE \code{datastack_name}. If unset a
sensible default is chosen.}

\item{live}{Whether to use live query mode, which updates any root ids to
their current value.}

\item{...}{Additional arguments passed to
\code{\link[fafbseg]{flywire_cave_query}}}
}
\value{
A data.frame
}
\description{
Query FANC tables in the CAVE annotation system
}
\examples{
\donttest{
nuclei_aug2021ver2=fanc_cave_query(table = "nuclei_aug2021ver2")
head(neuron_somas_dec2022)
hist(neuron_somas_dec2022$volume)
hist(neuron_somas_dec2022$volume^(1/3))
}
\dontrun{
points3d(xyzmatrix(nuclei_aug2021ver2$pt_position))
}
\dontrun{
library(dplyr)
cell_info=banc_cave_query('cell_info')
cell_info \%>\%
  filter(tag2=='anterior-posterior projection pattern') \%>\%
  count(tag)
}
}
\seealso{
\code{\link[fafbseg]{flywire_cave_query}}
}
\concept{fanc-cave}
