% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetta-api.R
\name{fanc_change_log}
\alias{fanc_change_log}
\title{Fetch change log information for one or more neurons}
\usage{
fanc_change_log(x, tz = "UTC", filtered = TRUE, OmitFailures = TRUE, ...)
}
\arguments{
\item{x}{One or more fanc ids in any format understandable by
\code{\link[fafbseg]{ngl_segments}}}

\item{tz}{Time zone for edit timestamps. Defaults to "UTC" i.e. Universal
Time, Coordinated. Set to "" for your current timezone. See
\code{\link{as.POSIXct}} for more details.}

\item{filtered}{Whether to filter out edits unlikely to relate to the current
state of the neuron (default \code{TRUE}, see details).}

\item{OmitFailures}{Whether to omit neurons for which there is an API timeout
or error. The default value (\code{TRUE}) will skip over errors, while
\code{NA}) will result in a hard stop on error. See \code{\link{nlapply}}
for more details.}

\item{...}{Additional arguments passed to \code{\link[fafbseg]{flywire_fetch}}}
}
\value{
a \code{data.frame} See \code{fabseg::\link{flywire_change_log}} for
details
}
\description{
Fetch change log information for one or more neurons
}
\details{
As of August 2021 this is a simple wrapper of
\code{fabseg::\link{flywire_change_log}}. For now the old (and less
convenient format) available from the zetta API can be obtained with the
private \code{fancr:::fanc_change_log_zetta} function.
}
\examples{
\donttest{
fanc_change_log("648518346473954669")
}
}
