% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners.R
\name{fanc_partner_summary}
\alias{fanc_partner_summary}
\title{Summarise the connectivity of FANC neurons}
\usage{
fanc_partner_summary(
  rootid,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0,
  datastack_name = NULL,
  ...
)
}
\arguments{
\item{rootid}{Character vector specifying one or more flywire rootids. As a
convenience for \code{flywire_partner_summary} this argument is passed to
\code{\link[fafbseg]{ngl_segments}} allowing you to pass in}

\item{partners}{Whether to fetch input or output synapses or both.}

\item{threshold}{For \code{flywire_partner_summary} only return partners with
greater than this number of connections to the query neuron(s) (default of
0 returns all connections)}

\item{remove_autapses}{For \code{flywire_partner_summary} whether to remove
autapses (defaults to TRUE)}

\item{cleft.threshold}{A threshold for the cleft score calculated by Buhmann
et al 2019 (default 0, we have used 30-100 to increase specificity)}

\item{datastack_name}{An optional CAVE datastack_name. If unset a sensible
default is chosen.}

\item{...}{Additional arguments passed to \code{\link{pbsapply}}}
}
\value{
a data.frame
}
\description{
Summarise the connectivity of FANC neurons
}
\details{
note that the rootids you pass in must be up to date. See example.
}
\examples{
head(fanc_partner_summary("648518346494405175"))
head(fanc_partner_summary("648518346494405175", partners='inputs'))
\dontrun{
# get the latest id for an outdate
fanc_partner_summary(fanc_latestid("648518346473954669"))

## open fanc/flywire scene containing top partners
library(dplyr)
fanc_partner_summary("648518346494405175", partners='inputs')
  slice_max(weight, n = 20) \%>\%
  purrr::pluck("post_id") \%>\%
  fanc_scene(open=TRUE)
}
}
\seealso{
\code{\link{flywire_partner_summary}}, \code{\link{fanc_latestid}}
}
