% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{fanc_scene}
\alias{fanc_scene}
\title{Return a sample Neuroglancer scene URL for FANC dataset}
\usage{
fanc_scene(ids = NULL, open = FALSE)
}
\arguments{
\item{ids}{A set of root ids to include in the scene. Can also be a
data.frame.}

\item{open}{Whether to open the URL in your browser (see
\code{\link{browseURL}})}
}
\value{
A character vector containing a single Neuroglancer URL (invisibly
when \code{open=TRUE}).
}
\description{
Return a sample Neuroglancer scene URL for FANC dataset
}
\details{
See
\href{https://fanc-reconstruction.slack.com/archives/C01RZP5JH9C/p1616522511001900}{FANC
slack} for details.
}
\examples{
\dontrun{
browseURL(fanc_scene())
fanc_scene(open=T)
fanc_scene("648518346498254576", open=T)
}
}
