% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{mirror_banc}
\alias{mirror_banc}
\title{Mirror points, neurons in BANC space}
\usage{
mirror_banc(x, units = c("nm", "microns", "raw"), subset = NULL)
}
\arguments{
\item{x}{Points, neurons or other objects compatible with \code{xyzmatrix}}

\item{units}{Units for both input \emph{and} output data.}

\item{subset}{Optional argument to transform only a subset of a neuron list.}
}
\value{
Trge transformed object (calibrated according to the units argument)
}
\description{
Mirror points, neurons in BANC space
}
\examples{
BANC.surf.m <- mirror_banc(BANC.surf)
\dontrun{
library(nat)
wire3d(BANC.surf)
# clearly not great in some places, especially optic lobe, but still useful
wire3d(BANC.surf.m, col='red')
}
}
